+++ lib/platforms/instagram.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+import 'social-media-generic.dart';
+
+class InstagramEmbedData extends SocialMediaGenericEmbedData {
+  final String embedHtml;
+
+  const InstagramEmbedData({@required this.embedHtml});
+
+  @override
+  String get htmlScriptUrl => 'https://www.instagram.com/embed.js';
+
+  @override
+  String get htmlBody => embedHtml + htmlScript;
+}