+++ lib/social_embed_webview.dart
@@ -1,65 +1,87 @@
 library social_embed_webview;
 
 import 'package:flutter/material.dart';
-import 'package:social_embed_webview/utlis.dart';
-import 'package:webview_flutter/webview_flutter.dart';
+import 'package:social_embed_webview/platforms/social-media-generic.dart';
+import 'package:social_embed_webview/utils/common-utils.dart';
 import 'package:url_launcher/url_launcher.dart';
-
-enum SocailMediaPlatforms {
-  twitter,
-  instagram,
-  facebook_post,
-  facebook_video,
-  youtube
-}
-
-final Map<SocailMediaPlatforms, String> _socailMediaScripts = {
-  SocailMediaPlatforms.twitter:
-      '<script src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>',
-  SocailMediaPlatforms.instagram:
-      '<script async src="https://www.instagram.com/embed.js"></script>',
-  SocailMediaPlatforms.youtube: '',
-  SocailMediaPlatforms.facebook_post:
-      '<script async defer src="https://connect.facebook.net/en_US/sdk.js#xfbml=1&version=v3.2"></script>',
-  SocailMediaPlatforms.facebook_video:
-      '<script async defer src="https://connect.facebook.net/en_US/sdk.js#xfbml=1&version=v3.2"></script>'
-};
+import 'package:webview_flutter/webview_flutter.dart';
 
 class SocialEmbed extends StatefulWidget {
-  final String embedCode;
-  final SocailMediaPlatforms type;
+  final SocialMediaGenericEmbedData socialMediaObj;
   final Color backgroundColor;
-
   const SocialEmbed(
-      {Key key,
-      @required this.embedCode,
-      @required this.type,
-      this.backgroundColor})
+      {Key key, @required this.socialMediaObj, this.backgroundColor})
       : super(key: key);
 
   @override
   _SocialEmbedState createState() => _SocialEmbedState();
 }
 
-class _SocialEmbedState extends State<SocialEmbed> {
+class _SocialEmbedState extends State<SocialEmbed> with WidgetsBindingObserver {
   double _height = 300;
+  WebViewController wbController;
+  SocialMediaGenericEmbedData smObj;
+  String htmlBody;
+
+  @override
+  void initState() {
+    super.initState();
+    smObj = widget.socialMediaObj;
+    htmlBody = getHtmlBody();
+    if (smObj.supportMediaControll) WidgetsBinding.instance.addObserver(this);
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    if (smObj.supportMediaControll) super.dispose();
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    switch (state) {
+      case AppLifecycleState.resumed:
+        break;
+      case AppLifecycleState.detached:
+        wbController.evaluateJavascript(smObj.stopVideoScript);
+        break;
+      case AppLifecycleState.inactive:
+      case AppLifecycleState.paused:
+        wbController.evaluateJavascript(smObj.pauseVideoScript);
+        break;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     final wv = WebView(
-        initialUrl: htmlToURI(_buildCode(context)),
+        initialUrl: htmlToURI(htmlBody),
         javascriptChannels:
             <JavascriptChannel>[_getHeightJavascriptChannel()].toSet(),
         javascriptMode: JavascriptMode.unrestricted,
-        navigationDelegate: (navigation) {
+        onPageStarted: (url) {
+          final color = colorToHtmlRGBA(getBackgroundColor(context));
+          wbController.evaluateJavascript(
+              'document.body.style= "background-color: $color"');
+        },
+        onWebViewCreated: (wbc) {
+          wbController = wbc;
+        },
+        onPageFinished: (str) {
+          if (smObj.aspectRatio == null)
+            wbController.evaluateJavascript('sendHeight()');
+        },
+        navigationDelegate: (navigation) async {
           final url = navigation.url;
-          canLaunch(url).then((can) => (can) ? launch(url) : null);
+          if (await canLaunch(url)) {
+            launch(url);
+          }
           return NavigationDecision.prevent;
         });
-    if (widget.type == SocailMediaPlatforms.youtube) {
-      return AspectRatio(aspectRatio: 16 / 9, child: wv);
-    }
-    return SizedBox(height: _height, child: wv);
+    final ar = smObj.aspectRatio;
+    return (ar != null)
+        ? AspectRatio(aspectRatio: ar, child: wv)
+        : SizedBox(height: _height, child: wv);
   }
 
   JavascriptChannel _getHeightJavascriptChannel() {
@@ -71,55 +93,36 @@ class _SocialEmbedState extends State<SocialEmbed> {
   }
 
   void _setHeight(double height) {
-    if (this.mounted) {
-      setState(() {
-        _height = height + 17;
-      });
-    }
-  }
-
-  String _buildCode(BuildContext context) {
-    return '<body style="background-color: ${colorToHtmlRGBA(getBackgroundColor(context))}"><div id="widget">${_parseData()}</div></body>' +
-        script +
-        _socailMediaScripts[widget.type];
-  }
-
-  String _parseData() {
-    if (widget.type == SocailMediaPlatforms.facebook_post) {
-      return '<div class="fb-post" data-href="${widget.embedCode}" data-show-text="true"></div>';
-    } else if (widget.type == SocailMediaPlatforms.facebook_video) {
-      return '<div class="fb-video" data-show-text="true" data-show-captions="true" data-href="${widget.embedCode}" data-show-text="true"></div>';
-    } else if (widget.type == SocailMediaPlatforms.youtube) {
-      return '<iframe src="https://www.youtube.com/embed/${widget.embedCode}" frameborder="0" allow="accelerometer;  encrypted-media; gyroscope; picture-in-picture" width=100% height=100%></iframe>';
-    }
-    return widget.embedCode;
+    setState(() {
+      _height = height + 17;
+    });
   }
 
   Color getBackgroundColor(BuildContext context) {
-    return (widget.backgroundColor == null)
-        ? Theme.of(context).scaffoldBackgroundColor
-        : widget.backgroundColor;
+    final color = widget.backgroundColor;
+    return (color == null) ? Theme.of(context).scaffoldBackgroundColor : color;
   }
-}
 
-String script = r"""
- <script type="text/javascript">
-	const elementHeightChangeListener = (ele,callback) => {
-		let lastHeight = ele.clientHeight;
-		return setInterval(() => {
-			const newHeight = ele.clientHeight;
-			if ( lastHeight != newHeight) {
-				callback(newHeight);
-				lastHeight = newHeight
-			}
-		},300);
-	}
-	
-	const widget = document.getElementById('widget');
-  PageHeight.postMessage(widget.clientHeight);
-  const interval =  elementHeightChangeListener(widget, (h) => {
-	  PageHeight.postMessage(h);
-	});
-	setInterval(() => clearTimeout(interval),10000);
-</script>
+  String getHtmlBody() => """
+      <body>
+        <div id="widget">${smObj.htmlBody}</div>
+        ${(smObj.aspectRatio == null) ? dynamicHeightScriptSetup : ''}
+        ${(smObj.canChangeSize) ? dynamicHeightScriptCheck : ''}
+      </body>
     """;
+
+  static const String dynamicHeightScriptSetup = """
+    <script type="text/javascript">
+      const widget = document.getElementById('widget');
+      const sendHeight = () => PageHeight.postMessage(widget.clientHeight);
+    </script>
+  """;
+
+  static const String dynamicHeightScriptCheck = """
+    <script type="text/javascript">
+      const onWidgetResize = (widgets) => sendHeight();
+      const resize_ob = new ResizeObserver(onWidgetResize);
+      resize_ob.observe(widget);
+    </script>
+  """;
+}