+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.3.0-dev.4] - Refactoring, Bug Fix and Dependencies update: 2020-12-11.
+* Breaking Changes (**Not backward compatiable**)
+* Bug Fix
+    * Webview auto dispose (Fixed by *webview_flutter*)
+    * Closing media content on widget dispose (Fixed by *webview_flutter*)
+    * Pausing/Stop media content on Widget State Change (limited to *Youtube Embed*)
+* **Remove Suport** of Facebook Video/Post (*Comming back soon*)
+* Dependencies Verion Update
+    * webview_flutter: *0.3.22+1* -> *1.0.7*
+    * url_launcher: *5.5.1* -> *5.7.10*
+
 ## [0.2.5] - Bug Fix: 2020-09-02.
 * Features
     * OnClick *native Open-In* option added